requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
bce-python-sdk==0.8.73
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
datasets==2.8.0
decorator==5.1.1
dill==0.3.4
easydict==1.10
entrypoints==0.4
fastapi==0.88.0
ffmpy==0.3.0
filelock==3.8.2
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gradio==3.15.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
huggingface-hub==0.11.1
idna==3.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.1
multidict==6.0.3
multiprocess==0.70.12.2
mysqlclient==2.1.1
numpy==1.24.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.8.3
packaging==22.0
paddle-bfloat==0.1.7
paddle2onnx==1.0.5
paddlefsl==1.1.0
paddlehub==2.3.1
paddlenlp==2.4.5
paddlepaddle==2.4.1
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.0
pyarrow==10.0.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pydub==0.25.1
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
rarfile==4.0
requests==2.28.1
responses==0.18.0
rfc3986==1.5.0
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
visualdl==2.4.1
websockets==10.4
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.2
